// src/app/shared/model/pages.model.ts
export interface pageSelection {
  skip: number;
  limit: number;
}
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface studentlist {
  sNo: number;
  id?: number;
  promotion?: string;
  result?: string;
  AdmissionNo?: string;
  isSelected: boolean;
  RollNo:string;
  name:string;
  class:string;
  section:string;
  gender:string;
  status:string;
  DateofJoin:string;
  DOB:string;
  imgSrc:string;
}